requirements.txt
Flask==3.0.0
google-generativeai
supabase
flask-cors
importlib-metadata>=4.6
zipp>=3.1.0
requests
python-dotenv
beautifulsoup4
lxml
Pillow
gunicorn
